requirements.txt
pyyaml
mysql-connector-python
azure-cosmos